ensure float

+++ geocat/comp/meteorology.py
@@ -1897,7 +1897,7 @@ def _delta_pressure1D(pressure_lev, surface_pressure):
         pressure_lev = np.flip(pressure_lev)
 
     # Calculate delta pressure
-    delta_pressure = np.full_like(pressure_lev, np.nan)
+    delta_pressure = np.full_like(pressure_lev, np.nan, dtype=float)
 
     # Determine which layers to calculate thickness for based upon midpoints
     midpoints = (pressure_lev + np.roll(pressure_lev, shift=1))/2
